add: project-manager-shim
