+++ app/gui/project-manager-shim-middleware/index.ts
@@ -2,7 +2,6 @@
  * @file A HTTP server middleware which handles routes normally proxied through to
  * the Project Manager.
  */
-import * as crypto from 'node:crypto'
 import * as fsSync from 'node:fs'
 import * as fs from 'node:fs/promises'
 import * as http from 'node:http'
@@ -10,9 +9,9 @@ import * as https from 'node:https'
 import * as path from 'node:path'
 
 import * as tar from 'tar'
-import * as yaml from 'yaml'
 
 import GLOBAL_CONFIG from 'enso-common/src/config.json' with { type: 'json' }
+import { handleFilesystemCommand } from 'project-manager-shim'
 
 import {
   AssetType,
@@ -70,72 +69,6 @@ const COOP_COEP_CORP_HEADERS = [
   ['Cross-Origin-Opener-Policy', 'same-origin'],
   ['Cross-Origin-Resource-Policy', 'same-origin'],
 ]
-
-// =============
-// === Types ===
-// =============
-
-/** Details of a project. */
-interface ProjectMetadata {
-  /** The name of the project. */
-  readonly name: string
-  /** The namespace of the project. */
-  readonly namespace: string
-  /** The project id. */
-  readonly id: string
-  /**
-   * The Enso Engine version to use for the project, represented by a semver version
-   * string.
-   *
-   * If the edition associated with the project could not be resolved, the
-   * engine version may be missing.
-   */
-  readonly engineVersion?: string
-  /** The project creation time. */
-  readonly created: string
-  /** The last opened datetime. */
-  readonly lastOpened?: string
-}
-
-/** Attributes of a file or folder. */
-interface Attributes {
-  readonly creationTime: string
-  readonly lastAccessTime: string
-  readonly lastModifiedTime: string
-  readonly byteSize: number
-}
-
-/** Metadata for an arbitrary file system entry. */
-type FileSystemEntry = DirectoryEntry | FileEntry | ProjectEntry
-
-/** The discriminator value for {@link FileSystemEntry}. */
-enum FileSystemEntryType {
-  DirectoryEntry = 'DirectoryEntry',
-  ProjectEntry = 'ProjectEntry',
-  FileEntry = 'FileEntry',
-}
-
-/** Metadata for a file. */
-interface FileEntry {
-  readonly type: FileSystemEntryType.FileEntry
-  readonly path: string
-  readonly attributes: Attributes
-}
-
-/** Metadata for a directory. */
-interface DirectoryEntry {
-  readonly type: FileSystemEntryType.DirectoryEntry
-  readonly path: string
-  readonly attributes: Attributes
-}
-
-/** Metadata for a project. */
-interface ProjectEntry {
-  readonly type: FileSystemEntryType.ProjectEntry
-  readonly path: string
-  readonly metadata: ProjectMetadata
-  readonly attributes: Attributes
-}
 
 // ====================================
 // === projectManagerShimMiddleware ===
@@ -305,176 +238,7 @@ export default function projectManagerShimMiddleware(
             .end('Command arguments must be an array of strings.')
         } else {
           void (async () => {
-            const toJSONRPCResult = (result: unknown) =>
-              JSON.stringify({ jsonrpc: '2.0', id: 0, result })
-            const toJSONRPCError = (message: string, data?: unknown) =>
-              JSON.stringify({
-                jsonrpc: '2.0',
-                id: 0,
-                error: { code: 0, message, ...(data != null ? { data } : {}) },
-              })
-            let result: string | fsSync.ReadStream = toJSONRPCError(
-              `Error running Project Manager command.`,
-              {
-                command: cliArguments,
-              },
-            )
-            try {
-              switch (cliArguments[0]) {
-                case '--filesystem-exists': {
-                  const directoryPath = cliArguments[1]
-                  if (directoryPath != null) {
-                    const exists = fsSync.existsSync(directoryPath)
-                    result = toJSONRPCResult({ exists })
-                  }
-                  break
-                }
-                case '--filesystem-list': {
-                  const directoryPath = cliArguments[1]
-                  if (directoryPath != null) {
-                    const entryNames = await fs.readdir(directoryPath)
-                    const entries: FileSystemEntry[] = []
-                    for (const entryName of entryNames) {
-                      const entryPath = path.join(directoryPath, entryName)
-                      if (isHidden(entryPath)) continue
-                      const stat = await fs.stat(entryPath)
-                      const attributes: Attributes = {
-                        byteSize: stat.size,
-                        creationTime: new Date(stat.ctimeMs).toISOString(),
-                        lastAccessTime: new Date(stat.atimeMs).toISOString(),
-                        lastModifiedTime: new Date(stat.mtimeMs).toISOString(),
-                      }
-                      if (stat.isFile()) {
-                        entries.push({
-                          type: FileSystemEntryType.FileEntry,
-                          path: entryPath,
-                          attributes,
-                        } satisfies FileEntry)
-                      } else {
-                        try {
-                          const packageMetadataPath = path.join(entryPath, 'package.yaml')
-                          const projectMetadataPath = path.join(
-                            entryPath,
-                            projectManagement.PROJECT_METADATA_RELATIVE_PATH,
-                          )
-                          const packageMetadataContents = await fs.readFile(packageMetadataPath)
-                          const packageMetadataYaml = yaml.parse(packageMetadataContents.toString())
-                          let projectMetadataJson
-                          try {
-                            const projectMetadataContents = await fs.readFile(projectMetadataPath)
-                            projectMetadataJson = JSON.parse(projectMetadataContents.toString())
-                          } catch (e) {
-                            if (
-                              'name' in packageMetadataYaml &&
-                              typeof packageMetadataYaml.name === 'string'
-                            ) {
-                              projectMetadataJson = {
-                                id: crypto.randomUUID(),
-                                kind: 'UserProject',
-                                created: new Date().toISOString(),
-                                lastOpened: null,
-                              }
-                              await fs.mkdir(path.dirname(projectMetadataPath), { recursive: true })
-                              await fs.writeFile(
-                                projectMetadataPath,
-                                JSON.stringify(projectMetadataJson),
-                              )
-                            } else {
-                              throw e
-                            }
-                          }
-                          const metadata = extractProjectMetadata(
-                            packageMetadataYaml,
-                            projectMetadataJson,
-                          )
-                          if (metadata != null) {
-                            // This is a project.
-                            entries.push({
-                              type: FileSystemEntryType.ProjectEntry,
-                              path: entryPath,
-                              attributes,
-                              metadata,
-                            } satisfies ProjectEntry)
-                          } else {
-                            // This error moves control flow to the
-                            // `catch` clause directly below.
-                            throw new Error('Invalid project metadata.')
-                          }
-                        } catch {
-                          // This is a regular directory, not a project.
-                          entries.push({
-                            type: FileSystemEntryType.DirectoryEntry,
-                            path: entryPath,
-                            attributes,
-                          } satisfies DirectoryEntry)
-                        }
-                      }
-                    }
-                    result = toJSONRPCResult({ entries })
-                  }
-                  break
-                }
-                case '--filesystem-create-directory': {
-                  const directoryPath = cliArguments[1]
-                  if (directoryPath != null) {
-                    await fs.mkdir(directoryPath, { recursive: true })
-                    result = toJSONRPCResult(null)
-                  }
-                  break
-                }
-                case '--filesystem-read-path': {
-                  const filePath = cliArguments[1]
-                  if (filePath != null) {
-                    result = await fsSync.createReadStream(filePath)
-                  }
-                  break
-                }
-                case '--filesystem-write-path': {
-                  const filePath = cliArguments[1]
-                  if (filePath != null) {
-                    await new Promise((resolve, reject) => {
-                      request
-                        .pipe(fsSync.createWriteStream(filePath), {
-                          end: true,
-                        })
-                        .on('close', resolve)
-                        .on('error', reject)
-                    })
-                    result = toJSONRPCResult(null)
-                  }
-                  break
-                }
-                case '--filesystem-move-from': {
-                  const sourcePath = cliArguments[1]
-                  const destinationPath = cliArguments[3]
-                  if (
-                    sourcePath != null &&
-                    cliArguments[2] === '--filesystem-move-to' &&
-                    destinationPath != null
-                  ) {
-                    await fs.rename(sourcePath, destinationPath)
-                    result = toJSONRPCResult(null)
-                  }
-                  break
-                }
-                case '--filesystem-delete': {
-                  const fileOrDirectoryPath = cliArguments[1]
-                  if (fileOrDirectoryPath != null) {
-                    await fs.rm(fileOrDirectoryPath, { recursive: true })
-                    result = toJSONRPCResult(null)
-                  }
-                  break
-                }
-                default: {
-                  const message = `Error in Project Manager shim: unknown command ${JSON.stringify(cliArguments)}`
-                  console.error(message)
-                  result = toJSONRPCError(message)
-                  break
-                }
-              }
-            } catch {
-              // Ignored. `result` retains its original value indicating an error.
-            }
+            const result = await handleFilesystemCommand(cliArguments, request, projectManagement)
 
             const resultData = typeof result === 'string' ? Buffer.from(result) : result
             if (resultData instanceof fsSync.ReadStream) {
@@ -557,61 +321,6 @@ export default function projectManagerShimMiddleware(
   } else {
     next()
   }
-}
-
-/**
- * Return a {@link ProjectMetadata} if the metadata is a valid metadata object,
- * else return `null`.
- */
-function extractProjectMetadata(yamlObj: unknown, jsonObj: unknown): ProjectMetadata | null {
-  if (
-    typeof yamlObj !== 'object' ||
-    yamlObj == null ||
-    typeof jsonObj !== 'object' ||
-    jsonObj == null
-  ) {
-    return null
-  } else {
-    const validDateString = (string: string) => {
-      const date = new Date(string)
-      return !Number.isNaN(Number(date)) ? date.toString() : null
-    }
-    const name = 'name' in yamlObj && typeof yamlObj.name === 'string' ? yamlObj.name : null
-    const namespace =
-      'namespace' in yamlObj && typeof yamlObj.namespace === 'string' ? yamlObj.namespace : 'local'
-    const engineVersion =
-      'edition' in yamlObj && typeof yamlObj.edition === 'string' ? yamlObj.edition : null
-    const id = 'id' in jsonObj && typeof jsonObj.id === 'string' ? jsonObj.id : null
-    const created =
-      'created' in jsonObj && typeof jsonObj.created === 'string' ?
-        validDateString(jsonObj.created)
-      : null
-    const lastOpened =
-      'lastOpened' in jsonObj && typeof jsonObj.lastOpened === 'string' ?
-        validDateString(jsonObj.lastOpened)
-      : null
-    if (name != null && id != null && created != null) {
-      return {
-        name,
-        namespace,
-        id,
-        ...(engineVersion != null ? { engineVersion } : {}),
-        created,
-        ...(lastOpened != null ? { lastOpened } : {}),
-      } satisfies ProjectMetadata
-    } else {
-      return null
-    }
-  }
-}
-
-/**
- * Checks if files that start with the dot.
- * Note on Windows does not check the hidden property.
- */
-function isHidden(filePath: string): boolean {
-  const dotfile = /(^|[\\/])\.[^\\/]+$/g
-  return dotfile.test(filePath)
 }
 
 /** Return whether a file exists. */
