+++ app/ide-desktop/client/package.json
@@ -27,6 +27,7 @@
     "mime-types": "^2.1.35",
     "mkcert": "3.2.0",
     "opener": "^1.5.2",
+    "project-manager-shim": "workspace:*",
     "semver": "^7.6.3",
     "string-length": "^5.0.1",
     "tar": "^6.2.1",
