+++ app/project-manager-shim/src/index.ts
@@ -1,0 +1,2 @@
+export * from './types.js'
+export * from './filesystem.js'
